Make boot up great again
Back when I was a boy, we had corny boot up screens
and we liked it. We'd walk up hill both ways just to turn
on our computers. It was always the dead of winter too,
and there was no sun; but we didn't complain.

Hell, even the narcoleptic os Windoze95 allowed
bored teenagers the ability to prank everyone in
the family with a funny image or message.